+++ src/components/dropdown/dropdown.js
@@ -3053,6 +3053,15 @@ Dropdown.prototype = {
       } else {
         this.select(option[0]);
         this.previousActiveDescendant = optionVal;
+
+        if (this.settings.noSearch && this.searchMatches) {
+          const char = option[0].innerText.toLowerCase().charAt(0);
+          Object.keys(this.searchMatches).forEach((key) => {
+            if (char !== key) {
+              this.searchMatches[key].length = 0;
+            }
+          });
+        }
       }
     } else {
       this.deselect(option[0]);